Got annotations back.
KiCad dropped my annotations from the oscillator when I used the
provided block copy and paste.  There seems to be no way to copy
a section of a schematic and keep the annotations.

As seems typical, the workaround is to do an end run around the
klunky Kicad Gui, copying files, writing one on top of the other, etc.

This is what people do.  This is what people have to do.  Aspects of
KiCad are awful, so amazingly bad that you wonder whether the CERN people
rescued the dullards who wrote KiCad, or instead maybe this is just good
enough for government work.  Really, it's embarrasing.